requirements.txt
Django
dj-database-url
mysqlclient
hashids
IPython[all]
djangorestframework
markdown
django-filter
